Moved remote search to SearchEndpoint
